Move logic from AudioVideo/Canonical container to AudioVideo/MediaPlayerSettings
AudioVideo/MediaPlayerSettings now takes in an audio or video block and returns the relevant settings object
